// apps/web/src/lib/hooks/schedule/types.ts
// Schedule class name enum values
export type ScheduleClassName =
    | 'Bugs'
    | 'Birds'
    | 'Beasts'
    | 'SuperBeasts'
    | 'FunnyBugs'
    | 'GiggleWorms'
    | 'GoodFriends'
    | 'FlipsHotshots'

// Schedule days enum values
export type ScheduleDay =
    | 'Monday'
    | 'Tuesday'
    | 'Wednesday'
    | 'Thursday'
    | 'Friday'
    | 'Saturday'
    | 'Sunday'